src/backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
pillow==9.1.1
python-multipart==0.0.5
numpy>=1.26.0,<2.1.0
